+++ tests/ZfcUserTest/Form/BaseTest.php
@@ -9,21 +9,9 @@ class BaseTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider providerTestConstruct
      */
-    public function testConstruct($useCaptcha = false)
+    public function testConstruct()
     {
-        $options = $this->getMock('ZfcUser\Options\RegistrationOptionsInterface');
-        $options->expects($this->once())
-                ->method('getUseRegistrationFormCaptcha')
-                ->will($this->returnValue($useCaptcha));
-        if ($useCaptcha && class_exists('\Zend\Captcha\AbstractAdapter')) {
-            $captcha = $this->getMockForAbstractClass('\Zend\Captcha\AbstractAdapter');
-
-            $options->expects($this->once())
-                    ->method('getFormCaptchaOptions')
-                    ->will($this->returnValue($captcha));
-        }
-
-        $form = new Form($options);
+        $form = new Form();
 
         $elements = $form->getElements();
 
@@ -36,11 +24,4 @@ class BaseTest extends \PHPUnit_Framework_TestCase
         $this->assertArrayHasKey('userId', $elements);
     }
 
-    public function providerTestConstruct()
-    {
-        return array(
-            array(true),
-            array(false)
-        );
-    }
 }
